Updated types and some other small changes
