+++ backend/src/controllers/preferencesController.ts
@@ -1,9 +1,14 @@
 import { Request, Response, NextFunction } from 'express';
 
-// import buildingData from '../interfaces/buildingData';
-import { getUserWithSubject } from './userController';
+import { getUserWithSubject, updatePreferences } from './userController';
 import * as responses from '../utils/responses';
+import BuildingData from '../types/buildingData';
+import Preferences from '../types/preferences';
 
+/**
+ * Gets current users preferences from DB
+ * @returns
+ */
 export const getPreferences = () => {
     const middleware = async (
         req: Request,
