--- backend/src/interfaces/buildingData.ts
@@ -1,6 +1,0 @@
-export default interface buildingData {
-    buildingId: string | null | undefined;
-    buildingName: string | null | undefined;
-    floorNames: string[] | null | undefined;
-    description: string | null | undefined;
-}
