+++ backend/src/routes/preferences.ts
@@ -1,6 +1,7 @@
 import express from 'express';
 import * as responses from '../utils/responses';
 import * as controller from '../controllers/preferencesController';
+import { validateBuildingInOrg } from '../controllers/buildingsController';
 
 export const router = express.Router();
