+++ backend/src/routes/rooms.ts
@@ -1,6 +1,7 @@
 import express from 'express';
 import { query } from 'express-validator';
 import * as controller from '../controllers/roomController';
+import { validateBuildingInOrg } from '../controllers/buildingsController';
 
 export const router = express.Router();
 
@@ -9,7 +10,7 @@ router.get(
     '/',
     query('showReserved').toBoolean(),
     query('building').trim().escape(),
-    controller.validateBuildingInOrg(),
+    validateBuildingInOrg(),
     controller.addAllRooms(),
     controller.fetchAvailability(),
     controller.writeReservationData(),
