+++ backend/src/types/buildingData.ts
@@ -1,0 +1,6 @@
+type BuildingData = {
+    id: string | null | undefined;
+    name: string | null | undefined;
+};
+
+export default BuildingData;
