+++ backend/src/types/preferences.ts
@@ -1,0 +1,7 @@
+import buildingData from './buildingData';
+
+type Preferences = {
+    building?: buildingData;
+};
+
+export default Preferences;
