+++ backend/src/types/roomData.ts
@@ -1,4 +1,4 @@
-export default interface roomData {
+type RoomData = {
     id: string | null | undefined;
     name: string | null | undefined;
     email: string | null | undefined;
@@ -7,4 +7,6 @@ export default interface roomData {
     floor: string | null | undefined;
     features: string[] | null | undefined;
     nextCalendarEvent: string | null | undefined;
-}
+};
+
+export default RoomData;
